+++ packages/ui-common/src/derivative/transformer.ts
@@ -1,0 +1,5 @@
+import { derivativeOrderTypeToGrpcOrderType } from './utils'
+
+export class DerivativeTransformer {
+  static derivativeOrderTypeToGrpcOrderType = derivativeOrderTypeToGrpcOrderType
+}