+++ packages/ui-common/src/spot/index.ts
@@ -1,5 +1,5 @@
+export * from './action'
 export * from './service'
 export * from './transformer'
-export * from './types'
 export * from './utils'
-export * from './data'
+export * from './types'