+++ packages/ui-common/src/spot/transformer.ts
@@ -1,0 +1,5 @@
+import { spotOrderTypeToGrpcOrderType } from './utils'
+
+export class SpotTransformer {
+  static spotOrderTypeToGrpcOrderType = spotOrderTypeToGrpcOrderType
+}