+++ packages/ui-common/src/spot/types.ts
@@ -1,0 +1,65 @@
+import {
+  SpotMarket as BaseUiSpotMarket,
+  SpotTrade,
+  SpotLimitOrder as UiSpotLimitOrder,
+  Orderbook as UiSpotOrderbook,
+  ChronosSpotMarketSummary,
+  SpotOrderSide,
+  AllChronosSpotMarketSummary,
+} from '@injectivelabs/spot-consumer'
+import { TradeDirection, TradeExecutionType } from '@injectivelabs/ts-types'
+import { Token } from '../token'
+import { Change, MarketBase, MarketType } from '../types'
+
+export interface UiBaseSpotMarket
+  extends Omit<BaseUiSpotMarket, 'quoteToken' | 'baseToken'> {}
+
+export interface UiBaseSpotMarketWithTokenMeta extends UiBaseSpotMarket {
+  quoteToken: Token
+  baseToken: Token
+}
+
+export interface UiSpotMarketWithTokenMeta
+  extends UiBaseSpotMarketWithTokenMeta {
+  priceDecimals: number
+  quantityDecimals: number
+  type: MarketType
+  subType: MarketType
+  marketBase?: MarketBase
+}
+
+export interface UiSpotTrade extends SpotTrade {
+  ticker?: string
+}
+
+export interface UiSpotMarketSummary extends ChronosSpotMarketSummary {
+  marketId: string
+  lastPrice?: number
+  lastPriceChange?: Change
+}
+
+export interface UiSpotMarketAndSummary {
+  marketId: string
+  summary: UiSpotMarketSummary
+}
+
+export enum SpotMarketMap {
+  UNSPECIFIED = 0,
+  BUY = 1,
+  SELL = 2,
+  STOP_BUY = 3,
+  STOP_SELL = 4,
+  TAKE_BUY = 5,
+  TAKE_SELL = 6,
+}
+
+export {
+  UiSpotLimitOrder,
+  TradeDirection,
+  SpotOrderSide,
+  TradeExecutionType,
+  BaseUiSpotMarket,
+  ChronosSpotMarketSummary,
+  AllChronosSpotMarketSummary,
+  UiSpotOrderbook,
+}