+++ packages/cosmos-ts/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.64](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/cosmos-ts@0.0.63...@injectivelabs/cosmos-ts@0.0.64) (2022-02-09)
+
+**Note:** Version bump only for package @injectivelabs/cosmos-ts
+
+
+
+
+
 ## [0.0.63](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/cosmos-ts@0.0.62...@injectivelabs/cosmos-ts@0.0.63) (2022-02-09)
 
 **Note:** Version bump only for package @injectivelabs/cosmos-ts