+++ packages/exchange-consumer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.137](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/exchange-consumer@0.0.136...@injectivelabs/exchange-consumer@0.0.137) (2022-02-09)
+
+**Note:** Version bump only for package @injectivelabs/exchange-consumer
+
+
+
+
+
 ## [0.0.136](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/exchange-consumer@0.0.135...@injectivelabs/exchange-consumer@0.0.136) (2022-02-09)
 
 **Note:** Version bump only for package @injectivelabs/exchange-consumer