+++ packages/ui-common/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.20](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/ui-common@0.0.19...@injectivelabs/ui-common@0.0.20) (2022-02-09)
+
+
+### Features
+
+* spot and derivative services and cleanup ([0705650](https://github.com/InjectiveLabs/injective-ts/commit/07056501c73d9d5218e32ec2483fdad2466d1c91))
+
+
+
+
+
 ## [0.0.19](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/ui-common@0.0.18...@injectivelabs/ui-common@0.0.19) (2022-02-09)